DEV: Move useful fixtures to common area
